requirements.txt
Pillow~=10.0.1
pyzbar~=0.1.9
requests~=2.31.0
pandas~=2.1.1
numpy~=1.26.0
streamlit~=1.27.0
opencv-python~=4.8.0.76
beautifulsoup4~=4.12.2